implement flattenTreeSystem
It turns out that we also need to filter by the supported system to get
a useful output out of flattenTree.

In order to keep back-compat, a new implementation has been created that
also looks at the meta.hydraPlatforms or meta.platforms if the former is
undefined, and only keep the derivations that have the provided system
in it.